+++ samples/counter/schema/src/main/kotlin/com/example/redwood/counter/schema.kt
@@ -16,7 +16,6 @@
 package com.example.redwood.counter
 
 import app.cash.redwood.layout.RedwoodLayout
-import app.cash.redwood.schema.Default
 import app.cash.redwood.schema.Property
 import app.cash.redwood.schema.Schema
 import app.cash.redwood.schema.Schema.Dependency
@@ -42,7 +41,6 @@ data class Text(
 data class Button(
   @Property(1) val text: String?,
   @Property(2)
-  @Default("true")
   val enabled: Boolean = true,
   @Property(3) val onClick: (() -> Unit)? = null,
 )